+++ modules/monitoring-function/locals.tf
@@ -5,6 +5,7 @@ locals {
     local.log_queries_appgw,
     local.log_queries_backup,
     local.log_queries_frontdoor,
+    local.log_queries_updates,
     local.log_queries_vpn,
   )
 }