+++ modules/monitoring-function/queries-vpn.tf
@@ -3,6 +3,7 @@ locals {
     vpn_successful_ike_diags = {
       MetricName = "fame.azure.virtual_network_gateway.ike_event_success"
       MetricType = "gauge"
+      QueryType  = "log_analytics"
       Query      = <<EOQ
         AzureDiagnostics
         | where Category == "IKEDiagnosticLog"
@@ -22,6 +23,7 @@ locals {
     vpn_tunnel_total_flow_count = {
       MetricName = "fame.azure.virtual_network_gateway.total_flow_count"
       MetricType = "gauge"
+      QueryType  = "log_analytics"
       Query      = <<EOQ
         AzureMetrics
         | where MetricName == "TunnelTotalFlowCount"